Feat : 리뷰 반영 및 지도 컴포넌트 분리 #75
