+++ next.config.js
@@ -14,7 +14,7 @@ module.exports = {
         protocol: 'https',
         hostname: 'sprint-fe-project.s3.ap-northeast-2.amazonaws.com',
         port: '',
-        pathname: '/globalnomad/**',
+        pathname: '/globalnomad/activity_registration_image/**',
       },
       {
         protocol: 'https',
