+++ src/components/Activities/MainContent/MainContent.module.css
@@ -26,15 +26,3 @@
   font-size: 1.6rem;
   line-height: 2.6rem;
 }
-
-.mapWrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 0.8rem;
-}
-
-.location {
-  display: flex;
-  align-items: center;
-  font-size: 1.4rem;
-}
