+++ src/components/Activities/ReviewList/ReviewList.module.css
@@ -53,6 +53,12 @@
   gap: 1.6rem;
 }
 
+.reviewWrapper > img {
+  border-radius: 100%;
+  flex-shrink: 0;
+  object-fit: cover;
+}
+
 .reviewerInfo {
   display: flex;
   flex-direction: column;
