+++ src/components/common/Map/Map.module.css
@@ -1,0 +1,11 @@
+.mapWrapper {
+  display: flex;
+  flex-direction: column;
+  gap: 0.8rem;
+}
+
+.location {
+  display: flex;
+  align-items: center;
+  font-size: 1.4rem;
+}
